// src/app/generated/model/workerResponse.ts
/**
 * BC Construction Services API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface WorkerResponse { 
    /**
     * Unique identifier of the worker
     */
    id?: number;
    /**
     * Worker\'s full name
     */
    name?: string;
    /**
     * Optional role/position
     */
    position?: string;
    /**
     * Current daily rate in PHP
     */
    dailyRate?: number;
    /**
     * Whether this worker is still active on the roster
     */
    active?: boolean;
    /**
     * ID of the user who added this worker
     */
    createdBy?: number;
    /**
     * Full name of the user who added this worker
     */
    createdByName?: string;
    /**
     * Timestamp when this worker was added
     */
    createdAt?: string;
    /**
     * Timestamp when this worker was last updated
     */
    updatedAt?: string;
}