// src/app/generated/model/workerUpdateRequest.ts
/**
 * BC Construction Services API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Request payload to update a worker\'s editable fields
 */
export interface WorkerUpdateRequest { 
    /**
     * Worker\'s full name
     */
    name: string;
    /**
     * Optional role/position
     */
    position?: string;
    /**
     * Daily rate in PHP
     */
    dailyRate: number;
}